+++ packages/adk-model-parser/lib/generator/generator.ts
@@ -1,6 +1,0 @@
-import chalk from 'chalk';
-import { Model } from '../types/types';
-
-export function generateModelFile(model: Model, fileName: string) {
-    console.log(chalk.green(`Generating model ${fileName} for ${model.Name}`));
-}