+++ packages/adk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.54.1 (2023-01-25)
+
+**Note:** Version bump only for package @codecatalyst/adk
+
+
+
+
+
+# 0.54.0 (2023-01-25)
+
+**Note:** Version bump only for package @codecatalyst/adk
+
+
+
+
+
 ## 0.53.1 (2023-01-24)
 
 **Note:** Version bump only for package @codecatalyst/adk